Codigo/backend/scripts/check_models.py
import google.generativeai as genai
import os
from dotenv import load_dotenv

load_dotenv()
api_key = os.getenv("GOOGLE_API_KEY")

if not api_key:
    print("❌ No encontré la API KEY.")
    exit()

genai.configure(api_key=api_key)

print("--- MODELOS DISPONIBLES PARA TI ---")
try:
    for m in genai.list_models():
        if 'generateContent' in m.supported_generation_methods:
            print(f"- {m.name}")
except Exception as e:
    print(f"❌ Error al conectar: {e}")